Moved apiTypes to api/types
