+++ package.json
@@ -20,7 +20,7 @@
     "@material-ui/icons": "^4.11.2",
     "@material-ui/lab": "^4.0.0-alpha.57",
     "@next/bundle-analyzer": "^11.0.1",
-    "@reduxjs/toolkit": "^1.6.0",
+    "@reduxjs/toolkit": "^1.6.1",
     "@sentry/nextjs": "^6.11.0",
     "@types/cookie": "^0.4.0",
     "@types/lodash.isempty": "^4.4.6",
@@ -40,7 +40,7 @@
     "next-redux-wrapper": "^6.0.2",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "react-redux": "^7.2.2",
+    "react-redux": "^7.2.5",
     "react-swipeable-views": "^0.14.0",
     "react-swipeable-views-utils": "^0.14.0",
     "redux-persist": "^6.0.0",
