+++ src/redux/saveLists/actions.ts
@@ -1,5 +1,5 @@
 import { createAction } from '@reduxjs/toolkit';
-import { Manga } from '../../common/apiTypes';
+import { Manga } from '../../api/types';
 import { ListType } from './types';
 
 export const favor = createAction<Manga>('saveLists/favor');
